+++ input/eingabe-disk.xml
@@ -8,7 +8,7 @@
         <particleContainerType>LC</particleContainerType>
         <r_cutoff>2.0</r_cutoff>
         <domainSizeX>20</domainSizeX>
-        <domainSizeY>15</domainSizeY>
+        <domainSizeY>20</domainSizeY>
         <domainSizeZ>20</domainSizeZ>
         <boundary1Type>2</boundary1Type>
         <boundary2Type>2</boundary2Type>
@@ -18,7 +18,7 @@
         <boundary6Type>2</boundary6Type>
         <particles>
             <x>10.0</x>
-            <y>7.5</y>
+            <y>10.0</y>
             <z>10.0</z>
             <velocityX>0.0</velocityX>
             <velocityY>-10.0</velocityY>
