+++ src/inputReader/FileReader.h
@@ -11,7 +11,7 @@
 
 #include <list>
 
-#include "../ParticleContainer.h"
+#include "../Container/ParticleContainer.h"
 /**
  * @brief FileReader class
  */
