+++ src/outputWriter/XYZWriter.h
@@ -12,7 +12,7 @@
 #include <fstream>
 #include <list>
 
-#include "ParticleContainer.h"
+#include "../Container/ParticleContainer.h"
 
 namespace outputWriter {
