+++ src/test/ParticleContainerTests.cpp
@@ -1,5 +1,5 @@
 #include "../Particle.h"
-#include "../ParticleContainer.h"
+#include "../Container/ParticleContainer.h"
 #include "spdlog/spdlog.h"
 #include <gtest/gtest.h>
